refactored locations
