+++ src/connection/mod.rs
@@ -1,6 +1,8 @@
 mod base;
 mod manager;
 
+pub mod location;
+
 pub use base::Connection;
 pub use manager::Manager;
