+++ src/connection/os/unix.rs
@@ -1,7 +1,6 @@
-use num_traits::FromPrimitive;
+use std::{env, net::Shutdown, os::unix::net::UnixStream, path::PathBuf};
 
 use crate::{connection::base::Connection, Result};
-use std::{env, net::Shutdown, os::unix::net::UnixStream, path::PathBuf};
 
 pub struct Socket {
     socket: UnixStream,
@@ -31,41 +30,6 @@ impl Connection for Socket {
 
     fn socket(&mut self) -> &mut Self::Socket {
         &mut self.socket
-    }
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash, num_derive::FromPrimitive)]
-#[repr(u8)]
-pub enum SocketLocation {
-    Root = 0b0000_0000,
-    Flatpak = 0b1000_0000,
-    Snap = 0b0100_0000,
-    SnapCanary = 0b1100_0000,
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub struct SocketId(u8);
-
-impl SocketId {
-    pub const fn new(number: u8, location: SocketLocation) -> Self {
-        let loc = location as u8;
-        Self(number | loc)
-    }
-
-    pub fn get_location(self) -> SocketLocation {
-        unsafe { self.try_get_location().unwrap_unchecked() }
-    }
-
-    pub fn try_get_location(self) -> Option<SocketLocation> {
-        SocketLocation::from_u8(self.0 & 0b1111_0000)
-    }
-
-    pub const fn get_number(self) -> u8 {
-        self.0 & 0b0000_1111
-    }
-
-    pub fn validate(self) -> bool {
-        self.get_number() < 10 && self.try_get_location().is_some()
     }
 }
